Refactor: structure RLayout
