--- src/components/Frame/Navigation.css
@@ -1,10 +1,0 @@
-#navigation {
-  position: fixed;
-  left: 0px;
-  bottom: 0px;
-  width: 100%;
-  height: 50px;
-  padding-bottom: calc(env(safe-area-inset-bottom));
-  box-shadow: 0 0 4px rgba(0, 0, 0, 0.1);
-  background-color: white;
-}
