+++ src/components/Frame/SubmitButton/SubmitButton.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useSpring, animated } from "react-spring";
 import PropTypes from "prop-types";
-import RLayout from "@frames/ReactiveLayout/RLayout";
+import RLayout from "components/common/RLayout";
 
 const Button = (props) => {
   const [isHover, setHover] = useState(false);
