+++ src/components/Mypage/PopupPolicy/PopupPolicy.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { animated, useSpring } from "react-spring";
 import { useHistory } from "react-router";
-import RLayout from "@frames/ReactiveLayout/RLayout";
+import RLayout from "components/common/RLayout";
 import PropTypes from "prop-types";
 import axios from "@tools/axios";
