+++ src/components/Skeleton/Footer.jsx
