+++ src/components/Skeleton/Skeleton.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { useLocation, Redirect } from "react-router-dom";
-import Navigation from "@frames/Navigation";
-import Footer from "@frames/Footer";
+import Navigation from "@components/Skeleton/Navigation";
+import Footer from "@components/Skeleton/Footer";
 import PopupPolicy from "@components/Mypage/PopupPolicy/PopupPolicy";
 import PropTypes from "prop-types";
 import axios from "@tools/axios";
