+++ tsconfig.paths.json
@@ -2,10 +2,8 @@
   "compilerOptions": {
     "baseUrl": "./src",
     "paths": {
-      "@/*": ["*"],
       "@components/*": ["components/*"],
       "@frames/*": ["components/Frame/*"],
-      "@assets/*": ["asset/*"],
       "@tools/*": ["tools/*"]
     }
   }
